Extension/chrome-extension/chatbot/backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
packaging>=24.1
function_schema>=0.4.5
celery==5.4.0
python-decouple==3.8
redis==5.0.7
qdrant-client==1.10.1
openai>=1.69.0
pymysql==1.1.1
mysql-connector-python
mysqlclient
SQLAlchemy
databases==0.9.0
llama-index-core<0.11.0,>=0.10.57
llama-index-llms-openai==0.1.31
llama-index>=0.10.56
langchain==0.2.14
langchain-community==0.2.12
cohere>=5.9.0
SPARQLWrapper==2.0.0
clickhouse-connect==0.8.15
flower==2.0.1
langchain_community
tqdm

# Thư viện cho các mô hình embedding mới
# scikit-learn>=1.0.2
# nltk>=3.7
# torch>=2.0.0
# transformers>=4.30.0
# rank-bm25>=0.2.2
# bmxpy>=0.2.0
# sentence-transformers>=2.2.2
# matplotlib>=3.5.0
# pandas>=1.3.5

scikit-learn
nltk
gensim
torch
transformers
rank-bm25
# bmxpy
baguetter
sentence-transformers
matplotlib
pandas

sentencepiece

# AI Summarization dependencies
transformers
torch
spacy
numpy